requirements.txt
Django==5.0.6
djangorestframework==3.15.1
joblib==1.4.2
gunicorn
django-cors-headers
numpy
dj-database-url
psycopg2-binary
psycopg2
scikit-learn